Insert product done

+++ src/main/java/com/example/javastoreapi/repositories/CategoryRepository.java
@@ -3,5 +3,5 @@ package com.example.javastoreapi.repositories;
 import com.example.javastoreapi.entities.Category;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoryRepository extends CrudRepository<Category, Long> {
+public interface CategoryRepository extends CrudRepository<Category, Byte> {
 }
